Bump minor versions of more core packages.
I missed these packages when publishing 1.6.2-beta.0, since I hadn't
bumped their versions in #9439 as I should have, so the release script did
not republish them. Because of this, 1.6.2-beta.0 will not be usable, and
so I will publish 1.6.2-beta.1 ASAP.

The boilerplate-generator version bump is patch rather than minor, since
the relevant changes to that package seemed backwards-compatible.